Config file doesn't need to export a function
This patch allows you to use the same format for the real config
file that you do for the defaults, just exporting an object with
a bunch of properties.  Any properties existing in defaults.js
and not in the config file supplied on the command line will be
copied in from defaults.

Also fix a couple console.err statements

Change-Id: I87e043baa020262668ee3be5108e30f1af202f48