Remove buttons for all widgets
Fixed to have unique modal id's

There is an issue with the modal backdrop so that is handled in the
reloading of the board until we figure out what to do about modals.

There is another issue with the fact that reloading the board triggers
the loading modal for x by y momentarily, which is ugly and bad.

Change-Id: I2087180c803b6934f7add6e2d2e5b0a3d3cf6a86