Use C3 binding in all widgets
Update C3 binding to delete chart on update with falsy value.
Would be nice to detect when we need to do this based on which
options have changed (e.g., not just data)

Modal IDs still need de-duplication to allow multiple instances of
the same widget type on one board.

Bug: T92430
Change-Id: I1e5b94d91efd2772a85b11fbceccc4688065e14f