Less complicated SQL for $/sec
Also more accurate, as it gives the total minutes elapsed this hour rather
than the number of minutes between the first and last donation this hour.

Still not sure if the metric is useful, but this change also lets it run in
strict SQL group by mode, the default on newer MariaDB versions.

Change-Id: I2d9d5bf786cb0b6d71f95889745c757b95a03d33