Rudiments of arbitrary campaigns
Only visible improvement is that it shows that this year's is 33
days. But it'll be a small step to add the year selector back and
get the campaign start days aligned.

FIXME: will blow up for campaigns that include New Years' Day. Also
the final day special case thing is still as wrong as parent commit.

Bug: T151820
Change-Id: Iaae238b2e29b80da47e54f2c206ca09baa3f53f8