Cache query promises rather than results
The idea is that if I make a query and you make the same query while mine
is still running (before there are any results to cache), we can both get
results as soon as the first query is finished. This will lighten the load
when for example lots of people are looking at the dash and the cache has
just run out.

Change-Id: If6b412a723d26f2f4dca229250fbf66cb54feb27