Add display values for groupings
Fixes display issue with "Is Recurring" and "Contribution Status"
filters when used as group-by columns. Contribution Status was
flat-out broken when there were statuses missing, because the
totals.sort() and counts.sort() got rid of the numeric indexes
referring to the different statuses. This patch maps numeric codes
to labels early in data processing, so the indexes are strings by
the time we get to the sort() functions.

Change-Id: Ie086a473020bf40ed2ce9f4ec3bad60c138fff2a