Add filters from wmf_contribution_extra
And disambiguate USD total from total in original currency.

This adds an optional join to wmf_contribution_extra for the x-by-y
and top 10 widgets, and makes it the preferred source for 'gateway',
'currency', and 'original amount' for those two. This should give us
better data for recurring donations, which don't generally have a
corresponding row in payments_initial.

Change-Id: Ia2da060c27ec653bf9b5ebb32f273290cf73ecf1